New realization option: No Poisson fluctuations
We added the option of populating the catalogue with the expected mean number of galaxies
(given the fluctuations of the density field) instead of with Poissonian realizations of
this mean density. This means that the catalog bins will show a float number (instead of
an int) that represents the number of galaxies without Poisson fluctuations.

Besides that we made minor changes to speed up the process: we stopped computing the
N_i catalogs and we pre-computed the cell volume.